// src/data/model/userSchema.ts
export interface UserModel {
  userId?: string;
  fullName?: string;
  username?: string;
  occupation?: string;
  uid: string;
  name: string | null;
  email: string | null;
  profilePic: string | null;
  socialType: 'facebook' | 'google' | 'phone';
  stepCompleted: number;
  notificationOn: boolean;
  matchOn: boolean;
  soundOn: boolean;
  online: boolean;
  status: string;
  role: string;
  DoB: string;
  photos: any;
  bio: string;
  height: string;
  bodyType: string;
  gender: string;
  sexuality: string;
  personality: string;
  education: string;
  maritalStatus: string;
  lookingFor: string;
  religion: string;
  drinkingStatus: string;
  smokingStatus: string;
  eatingStatus: string;
  createdAt: number;
  updatedAt: number;
  packageEndDate: any;
  dailySwipeCount: number;
  swipeStartDate: any;
  notificationReadCount: number;
  likedReadCount: number;
  seekerReadCount: number;
}